refactor(tui): organize charm renderer code
